🐛  We can't append nulls to FormData
When we were adding in the sub-Objects to the FormData, we were not
doing the same level of checks that we did at the top level - and that
was allowing `null` values to attempt to be added to the FormData -
which was blowing up. These chagnes now filter out the `null` values for
the FormData - as they are meant to be "nothing" anyway.