🔩  Standardized the Errors for all Calls
Patrick pointed out that it was a lot better to have a simple
convenience function to create a PostGridError from a string, and tag
the `type` with `client` as opposed to complicating the `errors` return
type from all the calls. This was super clear once we got into it. Much
better.

At the same time, we switched to the `>= 400` on the status code checks
so as not to have to worry about which of the `400`-series values it
was. This is just being smart, and again - thanks to Patrick for
pointing it out. 🙂