🧼  Moved the error structs to common ground
There was no reason to have the same error message defined in all the
different Api classes - so we moved them into the index.js, and then we
can import them as needed. This doesn't change the flow of the calls -
we still have to check on each call, but it makes updating the errors
easier, which is nice.