🔩  Previews now return PostalCode
The docs inticated that the output of the Address Previews was a
_truncated_ `postalOrZip`, but that's not the case... the complete
postal code is returned, and that's far more useful than the state of
province - as there's another function to get those from the postal
code. So we moved the client to pull the postal code. Yes, this could be
a parameter, or a state of the client, but we can add that later.