🐛  Add missing city on Contact
We were missing the `city` on the Contact data because it waas/is
missing from the PostGrid docs, and that's what I based it off. In their
examples, they do have the `city`, and that's what I assumed they meant
to have in the docs, so now it's added and tested. Good.