⚙️  Added Idempotency-Key support for Mail
In #15, the request to add support for the `Idempotency-Key` header for
the creation of Letters, Postcards, and Checks was something that needed
to be done. This is a way I've done it in the past where it's added as
an option to the end of the `create()` calls in each of the domains of
the mail client where it's applicable.